Permute names of modules
Fixing minor name permutation of modules.  Now module names make sense
given their functionality.
