+++ forestopenfermion/__init__.py
@@ -10,8 +10,10 @@
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 
-from forestopenfermion.pyquil_connector import exponentiate, TimeEvolution
-from forestopenfermion.pyquil_circuit_generator import (pyquilpauli_to_qubitop,
-                                                        qubitop_to_pyquilpauli)
+from forestopenfermion.pyquil_circuit_generator import (exponentiate,
+                                                        TimeEvolution)
+from forestopenfermion.pyquil_connector import (pyquilpauli_to_qubitop,
+                                                qubitop_to_pyquilpauli)
+
 
 __version__ = "0.0.2"
