+++ forestopenfermion/tests/pyquil_connector_test.py
@@ -7,8 +7,8 @@ from openfermion.ops import (FermionOperator, hermitian_conjugated,
                              QubitOperator, InteractionOperator,
                              InteractionRDM)
 from openfermion.transforms import jordan_wigner
-from forestopenfermion.pyquil_circuit_generator import (pyquilpauli_to_qubitop,
-                                                        qubitop_to_pyquilpauli)
+from forestopenfermion.pyquil_connector import (pyquilpauli_to_qubitop,
+                                                qubitop_to_pyquilpauli)
 from pyquil.paulis import PauliTerm, PauliSum
